# backend/typescript/swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: UW Blueprint Starter Code TypeScript API
  description: An API for the TypeScript Backend of Starter Code


# Authentication 
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
...
paths:
...
  /users:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            # type: integer (postgres)
            # minimum: 1
            type: string
          description: The user ID
        - in: query
          name: email
          schema:
            type: string
          description: The user's email
      tags:
        - User
      description: Get all users, optionally filter by a userId or email query parameter to retrieve a single user
      responses:
        '200':
          description: Successfully returned a list of users, or a user filtered by id or email
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                      - id, firstName, lastName, email, role, phoneNumber
                  properties:
                    id:
                      type: string
                    firstName:
                      type: string
                    lastName:
                      type: string
                    email:
                      type: string
                    role:
                      type: string
                    phoneNumber:
                      type: string
        '400':
          description: Cannot query by both user and email
        '401':
          description: Unauthorized
        '500':
          description: Invalid request


    post:
      security:
        - bearerAuth: []
      tags:
        - User
      description: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName, lastName, role, email, phoneNumber
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                role:
                  type: string
                  enum: ["User", "Admin", "Volunteer", "Donor"]
                email:
                  type: string
                phoneNumber:
                  type: string

      responses:
        '201':
          description: Successfully created a new user
          content:
            application/json:
              schema:
                type: object
                required:
                    - id, firstName, lastName, email, role, phoneNumber
                properties:
                  id:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  phoneNumber:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Invalid request

          
    delete:
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            # type: integer (postgres)
            # minimum: 1
            type: string
          description: The user ID
        - in: query
          name: email
          schema:
            type: string
          description: The user's email
      tags:
        - User
      description: Deletes a user by userId or email, specified through a query parameter
      responses:
        '204':
          description: Succesfully deleted user
        '400':
          description: Cannot query by both userId and email, invalid userId or email, userId or email is not a string, userId or email not supplied
        '401':
          description: Unauthorized
        '500':
          description: Invalid request
  /users/{userId}:
    put:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            # type: integer (postgres)
            # minimum: 1
            type: string
          description: The user ID
      tags:
        - User
      description: Update the user with the specified userId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName, lastName, role, email, phoneNumber
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                role:
                  type: string
                  enum: ["User", "Admin"]
                email:
                  type: string
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Successfully updated user
          content:
            application/json:
              schema:
                type: object
                required:
                    - id, firstName, lastName, email, role, phoneNumber
                properties:
                  id:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  phoneNumber:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Invalid request